Add privacy checkbox
